refactor: query 훅 분리 및 onSuccess 콜백 옵션 제거
